+++ src/kio/schema/alter_partition/request/v2.py
@@ -7,20 +7,23 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import BrokerId
+from kio.schema.primitive import i8
+from kio.schema.primitive import i32
+from kio.schema.primitive import i64
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class PartitionData:
     __flexible__: ClassVar[bool] = True
-    partition_index: int = field(metadata={"kafka_type": "int32"})
+    partition_index: i32 = field(metadata={"kafka_type": "int32"})
     """The partition index"""
-    leader_epoch: int = field(metadata={"kafka_type": "int32"})
+    leader_epoch: i32 = field(metadata={"kafka_type": "int32"})
     """The leader epoch of this partition"""
     new_isr: tuple[BrokerId, ...] = field(metadata={"kafka_type": "int32"}, default=())
     """The ISR for this partition"""
-    leader_recovery_state: int = field(metadata={"kafka_type": "int8"}, default=0)
+    leader_recovery_state: i8 = field(metadata={"kafka_type": "int8"}, default=i8(0))
     """1 if the partition is recovering from an unclean leader election; 0 otherwise."""
-    partition_epoch: int = field(metadata={"kafka_type": "int32"})
+    partition_epoch: i32 = field(metadata={"kafka_type": "int32"})
     """The expected epoch of the partition which is being updated. For legacy cluster this is the ZkVersion in the LeaderAndIsr request."""
 
 
@@ -37,6 +40,6 @@ class AlterPartitionRequest:
     __flexible__: ClassVar[bool] = True
     broker_id: BrokerId = field(metadata={"kafka_type": "int32"})
     """The ID of the requesting broker"""
-    broker_epoch: int = field(metadata={"kafka_type": "int64"}, default=-1)
+    broker_epoch: i64 = field(metadata={"kafka_type": "int64"}, default=i64(-1))
     """The epoch of the requesting broker"""
     topics: tuple[TopicData, ...]
