+++ src/kio/schema/alter_partition/response/v2.py
@@ -7,24 +7,27 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import BrokerId
+from kio.schema.primitive import i8
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class PartitionData:
     __flexible__: ClassVar[bool] = True
-    partition_index: int = field(metadata={"kafka_type": "int32"})
+    partition_index: i32 = field(metadata={"kafka_type": "int32"})
     """The partition index"""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The partition level error code"""
     leader_id: BrokerId = field(metadata={"kafka_type": "int32"})
     """The broker ID of the leader."""
-    leader_epoch: int = field(metadata={"kafka_type": "int32"})
+    leader_epoch: i32 = field(metadata={"kafka_type": "int32"})
     """The leader epoch."""
     isr: tuple[BrokerId, ...] = field(metadata={"kafka_type": "int32"}, default=())
     """The in-sync replica IDs."""
-    leader_recovery_state: int = field(metadata={"kafka_type": "int8"}, default=0)
+    leader_recovery_state: i8 = field(metadata={"kafka_type": "int8"}, default=i8(0))
     """1 if the partition is recovering from an unclean leader election; 0 otherwise."""
-    partition_epoch: int = field(metadata={"kafka_type": "int32"})
+    partition_epoch: i32 = field(metadata={"kafka_type": "int32"})
     """The current epoch for the partition for KRaft controllers. The current ZK version for the legacy controllers."""
 
 
@@ -39,8 +42,8 @@ class TopicData:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AlterPartitionResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The top level response error code"""
     topics: tuple[TopicData, ...]
