+++ src/kio/schema/describe_client_quotas/request/v0.py
@@ -5,13 +5,15 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i8
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class ComponentData:
     __flexible__: ClassVar[bool] = False
     entity_type: str = field(metadata={"kafka_type": "string"})
     """The entity type that the filter component applies to."""
-    match_type: int = field(metadata={"kafka_type": "int8"})
+    match_type: i8 = field(metadata={"kafka_type": "int8"})
     """How to match the entity {0 = exact name, 1 = default name, 2 = any specified name}."""
     match: str | None = field(metadata={"kafka_type": "string"})
     """The string to match against, or null if unused for the match type."""
